+++ CHANGELOG.md
@@ -27,6 +27,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [release-009] - 2022-03-11
 
+### Added
+
+- Add accessibility statement
+
 ### Changed
 
 - Fix wrong links on dashboard